requirements.txt
discord.py
os
discord.ext
python-dotenv